+++ grails-app/services/org/pih/warehouse/importer/ProductAttributeImportDataService.groovy
@@ -10,19 +10,19 @@
 package org.pih.warehouse.importer
 
 import grails.gorm.transactions.Transactional
-import org.pih.warehouse.core.UnitOfMeasure
-import org.pih.warehouse.product.Attribute
-import org.pih.warehouse.product.Product
 import org.pih.warehouse.product.ProductAttribute
+import org.pih.warehouse.product.ProductService
 import org.springframework.validation.BeanPropertyBindingResult
 
 
 @Transactional
 class ProductAttributeImportDataService implements ImportDataService {
+    ProductService productService
 
+    @Override
     void validateData(ImportDataCommand command) {
         command.data.eachWithIndex { params, index ->
-            ProductAttribute productAttribute = createOrUpdateProductAttribute(params)
+            ProductAttribute productAttribute = productService.createOrUpdateProductAttribute(params)
             if (!productAttribute.validate()) {
                 productAttribute.errors.each { BeanPropertyBindingResult error ->
                     command.errors.reject("Row ${index + 1}: Product attribute ${productAttribute} is invalid: ${error.getFieldError()}")
@@ -31,27 +31,15 @@ class ProductAttributeImportDataService implements ImportDataService {
         }
     }
 
+    @Override
     void importData(ImportDataCommand command) {
         command.data.eachWithIndex { params, index ->
-            ProductAttribute productAttribute = createOrUpdateProductAttribute(params)
+            ProductAttribute productAttribute = productService.createOrUpdateProductAttribute(params)
             if (productAttribute.validate()) {
                 productAttribute.product.save(failOnError: true)
             }
         }
     }
 
-    ProductAttribute createOrUpdateProductAttribute(Map params) {
-        Product product = Product.findByProductCode(params.productCode)
-        Attribute attribute = Attribute.findByCode(params.attributeCode)
-        UnitOfMeasure unitOfMeasure = UnitOfMeasure.findByCode(params.unitOfMeasureCode)
-        ProductAttribute productAttribute = ProductAttribute.findByProductAndAttribute(product, attribute)
-        if (!productAttribute) {
-            productAttribute = new ProductAttribute()
-            productAttribute.attribute = attribute
-            product.addToAttributes(productAttribute)
-        }
-        productAttribute.value = params.attributeValue
-        productAttribute.unitOfMeasure = unitOfMeasure?:attribute?.unitOfMeasureClass?.baseUom
-        return productAttribute
-    }
+
 }
